Feat: migration channels page
