+++ components/common/scheduleCard/ScheduleCard.tsx
@@ -1,5 +1,4 @@
 'use client';
-/* eslint-disable react-hooks/exhaustive-deps */
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { ContentsDataType } from '@/types/inSheet';
 import Link from 'next/link';
