+++ components/layout/Header.tsx
@@ -39,7 +39,6 @@ const Header = () => {
     return () => {
       window.removeEventListener('scroll', handleScroll);
     };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
