--- pages/_document.tsx
@@ -1,15 +1,0 @@
-import { Html, Head, Main, NextScript } from 'next/document';
-
-const Document = () => {
-  return (
-    <Html lang="ko">
-      <Head />
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-};
-
-export default Document;
